feat: Implement job application quiz and interview management features with dedicated frontend pages and backend services.

+++ services/user/src/controllers/user.ts
@@ -85,7 +85,7 @@ export const updateProfilePic = TryCatch(
       throw new ErrorHandler(500, "failed to generate buffer");
     }
 
-    const { data: uploadResult } : any = await axios.post(
+    const { data: uploadResult }: any = await axios.post(
       `${process.env.UPLOAD_SERVICE}/api/utils/upload`,
       {
         buffer: fileBuffer.content,
@@ -93,7 +93,7 @@ export const updateProfilePic = TryCatch(
       }
     );
 
-    const [updatedUser]  = await sql`
+    const [updatedUser] = await sql`
     UPDATE users SET profile_pic = ${uploadResult.url}, profile_pic_public_id = ${uploadResult.public_id} WHERE user_id = ${user.user_id} RETURNING user_id, name, profile_pic;
     `;
 
@@ -125,7 +125,7 @@ export const updateResume = TryCatch(async (req: AuthenticatedRequest, res) => {
     throw new ErrorHandler(500, "failed to generate buffer");
   }
 
-  const { data: uploadResult } : any = await axios.post(
+  const { data: uploadResult }: any = await axios.post(
     `${process.env.UPLOAD_SERVICE}/api/utils/upload`,
     {
       buffer: fileBuffer.content,
@@ -208,9 +208,8 @@ export const deleteSkillFromUser = TryCatch(
       throw new ErrorHandler(400, "Please provide a skill name");
     }
 
-    const result = await sql`DELETE FROM user_skills WHERE user_id = ${
-      user.user_id
-    } AND skill_id = (SELECT skill_id FROM skills WHERE name = ${skillName.trim()}) RETURNING user_id;`;
+    const result = await sql`DELETE FROM user_skills WHERE user_id = ${user.user_id
+      } AND skill_id = (SELECT skill_id FROM skills WHERE name = ${skillName.trim()}) RETURNING user_id;`;
 
     if (result.length === 0) {
       throw new ErrorHandler(404, `Skill ${skillName.trim()} was not found`);
@@ -289,7 +288,11 @@ export const applyForJob = TryCatch(async (req: AuthenticatedRequest, res) => {
 export const getAllaplications = TryCatch(
   async (req: AuthenticatedRequest, res) => {
     const applications = await sql`
-    SELECT a.*, j.title AS job_title, j.salary AS job_salary, j.location AS job_location FROM applications a JOIN jobs j ON a.job_id = j.job_id WHERE a.applicant_id = ${req.user?.user_id}
+    SELECT a.*, j.title AS job_title, j.salary AS job_salary, j.location AS job_location, i.meet_link, i.scheduled_at 
+    FROM applications a 
+    JOIN jobs j ON a.job_id = j.job_id 
+    LEFT JOIN interviews i ON a.application_id = i.application_id
+    WHERE a.applicant_id = ${req.user?.user_id}
   `;
 
     res.json(applications);
